moved AverageTerrain to its own file
